requirements.txt
django
hypothesis
faker
pytz
djangorestframework
markdown
django-filter
beautifulsoup4
coverage
